MOTECH-2093: renamed mds.web.ex to mds.web.exception for consistency
