+++ platform/mds/mds-web/src/main/java/org/motechproject/mds/web/rest/MdsRestController.java
@@ -12,7 +12,7 @@ import org.motechproject.mds.ex.rest.RestNotSupportedException;
 import org.motechproject.mds.ex.rest.RestOperationNotSupportedException;
 import org.motechproject.mds.query.QueryParams;
 import org.motechproject.mds.rest.MdsRestFacade;
-import org.motechproject.mds.web.ex.InvalidParameterException;
+import org.motechproject.mds.web.exception.InvalidParameterException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
